requirements.txt
Flask==3.1.3
psycopg2==2.9.12
python-dotenv==1.2.2
argon2-cffi==25.1.0
Flask-Limiter==4.1.1
Flask-WTF==1.3.0